completed notifications page

+++ lib/repositories/notifications_repository.dart
@@ -1,16 +1,34 @@
 import 'package:altar_of_prayers/database/notifications_dao.dart';
+import 'package:altar_of_prayers/graphql/graphql.dart';
 import 'package:altar_of_prayers/models/notification.dart';
 import 'package:altar_of_prayers/repositories/user_repository.dart';
+import 'package:graphql_flutter/graphql_flutter.dart';
 
 class NotificationsRepository {
   NotificationsDao _notificationsDao = NotificationsDao();
   UserRepository _userRepository = UserRepository();
+  GraphQLConfiguration graphQLConfiguration = GraphQLConfiguration();
+  QueryMutation queryMutation = QueryMutation();
 
   Future<List<Map<String, dynamic>>> getLocalNotifications() async {
     List<Map<String, dynamic>> notifications =
         await _notificationsDao.getNotifications();
     if (notifications != null) return notifications;
     return [];
+  }
+
+  Future<bool> markNotificationAsRead({int id}) async {
+    GraphQLClient _client = await graphQLConfiguration.clientToQuery();
+    QueryResult result = await _client.mutate(
+      MutationOptions(
+          documentNode: gql(queryMutation.markNotificationAsRead(id: id))),
+    );
+
+    if (!result.hasException) {
+      return result.data["markNotificationAsRead"]["success"];
+    } else {
+      return false;
+    }
   }
 
   Future<List> getNotificationsFromServer() async {
@@ -20,17 +38,21 @@ class NotificationsRepository {
 
     _notificationsDao.deleteAllNotifications();
 
-    List<Future<NotificationModel>> _userNotifications =
+    List<NotificationModel> _userNotifications =
         (currentUserInfo["currentUser"]["userNotification"] as List)
-            .map((notification) async {
-      NotificationModel _notification = NotificationModel.fromServerJson(notification);
+            .map((notification) {
+      NotificationModel _notification =
+          NotificationModel.fromServerJson(notification);
 
-      // save notification and return
-      await _notificationsDao.saveNotification(notification: _notification);
       return _notification;
     }).toList();
 
-
+    Future.forEach(_userNotifications, (notification) async {
+      // save notification and return
+      // print("notification: ${notification.runtimeType}");
+      await _notificationsDao.saveNotification(notification: notification);
+      return null;
+    });
     return _userNotifications;
   }
 
@@ -39,6 +61,18 @@ class NotificationsRepository {
   }
 
   Future deleteNotification({NotificationModel notification}) async {
-    return _notificationsDao.deleteNotification(notification: notification);
+    GraphQLClient _client = await graphQLConfiguration.clientToQuery();
+    QueryResult result = await _client.mutate(
+      MutationOptions(
+          documentNode: gql(queryMutation.deleteUserNotification(id: notification.id))),
+    );
+
+    if (!result.hasException) {
+      if (result.data["deleteUserNotification"]["success"])
+        await _notificationsDao.deleteNotification(notification: notification);
+      return result.data["deleteUserNotification"]["success"];
+    } else {
+      return false;
+    }
   }
 }
